WCM-982: Extract docker test layer basics, so we can reuse them
